Add Logging Process and utilities.
The design of this code is to let NewLogProcess() create a pointer to a
LogProcess instance which acts as the API proxy for the go-routine that
deals with the actual writing of the messages to the log file.

The LogProcess.Log() method lets you log a string to the log file and
its sister method, LogProcess.LogF(), lets you log a format string to a
file.

The LogProcess.Close() method allows you to shutdown the logging
go-routine and mark the LogProcess as being invalid. Any calls to an
invalid LogProcess is considered to be a void operation for now.

Later we might end up making API calls to an invalid LogProcess as a
fatal operation instead of void, but for now it seems to be good enough.