Add ClientSet implementation.
This patch implement a data-structure to represent a set of unique
client's. It's currently implemented as a hack where we (ab)use Go's
build-in, type-safe, map data-type to represent the underlying set.

At some point in the future, if Go ends up gaining support for a simple
form of parametric types, it would be nice to reimplement this code
using a proper data-structure (either as a hash-set or using a
self-balanced binary tree).

The Find() method have been left out, since it's most likely not going
to be missed.