Let NewClient() handle RegisteredProtocolHandler() instead of returning.
The current implementation of NewClient() returns as soon as the given
client moves from being an unregistered client to a registeret client.
This means that ConnectionHandler(), which is shared between each
listener-type (TCP and WebSocket), will return as well. If the
ConnectionHandler() function returns, the Go HTTP package, which handles
WebSocket connections, will effectively terminate the WebSocket
connection even though the connection is actually still to be kept
alive.

Fixes #17